Restructure the preferences panel, moving it into its own package to resolve circular import issues
