+++ src/sas/qtgui/Utilities/Preferences/PreferencesWidget.py
@@ -1,12 +1,9 @@
-import logging
+import functools
 
-from PyQt5.QtWidgets import QComboBox, QDialog, QPushButton, QWidget, QLabel, QHBoxLayout, QVBoxLayout, QLineEdit, QCheckBox
-from typing import Optional, Any, List, Union, Callable
+from PyQt5.QtWidgets import QComboBox, QWidget, QLabel, QHBoxLayout, QVBoxLayout, QLineEdit, QCheckBox
+from typing import Optional, List, Union, Callable, Any
 
-from sas.system.config.config import config
-from sas.qtgui.Utilities.UI.PreferencesUI import Ui_preferencesUI
-
-logger = logging.getLogger(__name__)
+from sas.system import config
 
 
 def set_config_value(attr: str, value: Any):
@@ -37,69 +34,15 @@ def cb_replace_all_items_with_new(cb: QComboBox, new_items: List[str], default_i
     cb.setCurrentIndex(index)
 
 
-class PreferencesPanel(QDialog, Ui_preferencesUI):
-    """A preferences panel to house all SasView related settings. The left side of the window is a listWidget with a
-    options menus available. The right side of the window is a stackedWidget object that houses the options
-    associated with each listWidget item.
-    **Important Note** When adding new preference widgets, the index for the listWidget and stackedWidget *must* match
-    Release notes:
-    SasView v5.0.5: Added defaults for loaded data units and plotted units
+def config_value_setter_generator(attr: str,dtype: Optional[Any] = None):
+    """Helper method that generates a callback to set a config value.
+
+    :param attr: name of the attribute to set
+    :param dtype:
+    :return: a function that takes a single argument, which will be cast to dtype
+            and set in config as attr
     """
-
-    def __init__(self, parent=None):
-        super(PreferencesPanel, self).__init__(parent)
-        self.setupUi(self)
-        self.parent = parent
-        self.setWindowTitle("Preferences")
-        self.warning = None
-        # A list of callables used to restore the default values for each item in StackedWidget
-        self.restoreDefaultMethods = []
-        # Set defaults values for the list and stacked widgets
-        self.stackedWidget.setCurrentIndex(0)
-        self.listWidget.setCurrentRow(0)
-        # Add window actions
-        self.listWidget.currentItemChanged.connect(self.prefMenuChanged)
-        self.buttonBox.clicked.connect(self.onClick)
-
-    def prefMenuChanged(self):
-        """When the preferences menu selection changes, change to the appropriate preferences widget """
-        row = self.listWidget.currentRow()
-        self.stackedWidget.setCurrentIndex(row)
-
-    def onClick(self, btn: QPushButton):
-        """Handle button click events in one area"""
-        # Reset to the default preferences
-        if btn.text() == 'Restore Defaults':
-            self.restoreDefaultPreferences()
-        elif btn.text() == 'OK':
-            self.close()
-        elif btn.text() == 'Help':
-            self.help()
-
-    def restoreDefaultPreferences(self):
-        """Reset all preferences to their default preferences"""
-        for method in self.restoreDefaultMethods:
-            if callable(method):
-                method()
-            else:
-                logger.warning(f'While restoring defaults, {str(method)} of type {type(method)}'
-                               + ' was given. A callable object was expected.')
-
-    def close(self):
-        """Save the configuration values when the preferences window is closed"""
-        config.save()
-        super(PreferencesPanel, self).close()
-
-    def addWidget(self, widget: QWidget):
-        self.stackedWidget.addWidget(widget)
-        self.listWidget.addItem(widget.name)
-        if widget.resetDefaults is not None and callable(widget.resetDefaults):
-            self.restoreDefaultMethods.append(widget.resetDefaults)
-
-    def help(self):
-        """Open the help window associated with the preferences window"""
-        tree_location = "/user/qtgui/MainWindow/preferences_help.html"
-        self.parent.showHelp(tree_location)
+    return functools.partial(set_config_value,attr=attr,dtype=dtype)
 
 
 class PreferencesWidget(QWidget):
@@ -112,8 +55,8 @@ class PreferencesWidget(QWidget):
         super(PreferencesWidget, self).__init__()
         self.name = name
         self.resetDefaults = default_method
-        self.horizontalLayout = QHBoxLayout()
-        self.setLayout(self.horizontalLayout)
+        self.verticalLayout = QHBoxLayout()
+        self.setLayout(self.verticalLayout)
         self.adjustSize()
 
     def _createLayoutAndTitle(self, title: str):
@@ -121,7 +64,7 @@ class PreferencesWidget(QWidget):
         :param title: The title of the interactive item to be added to the preferences panel.
         :return: A QVBoxLayout instance with a title box already added
         """
-        layout = QVBoxLayout(self)
+        layout = QHBoxLayout()
         label = QLabel(title + ": ", self)
         layout.addWidget(label)
         return layout
@@ -139,7 +82,7 @@ class PreferencesWidget(QWidget):
         cb_replace_all_items_with_new(box, params, default)
         box.currentIndexChanged.connect(callback)
         layout.addWidget(box)
-        self.horizontalLayout.addLayout(layout)
+        self.verticalLayout.addLayout(layout)
 
     def addTextInput(self, title: str, callback: Callable, default_text: Optional[str] = ""):
         """Add a title and text box within the widget.
@@ -153,7 +96,7 @@ class PreferencesWidget(QWidget):
             text_box.setText(default_text)
         text_box.textChanged.connect(callback)
         layout.addWidget(text_box)
-        self.horizontalLayout.addLayout(layout)
+        self.verticalLayout.addLayout(layout)
 
     def addCheckBox(self, title: str, callback: Callable, checked: Optional[bool] = False):
         """Add a title and check box within the widget.
@@ -166,4 +109,4 @@ class PreferencesWidget(QWidget):
         check_box.setChecked(checked)
         check_box.toggled.connect(callback)
         layout.addWidget(check_box)
-        self.horizontalLayout.addLayout(layout)
+        self.verticalLayout.addLayout(layout)
